requirements.txt
streamlit
pandas
docxtpl
google-generativeai
openpyxl
python-docx